+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./lib/date";
 export * from "./lib/enum";
 export * from "./lib/localStorage";
+export * from "./lib/string";