+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./lib/date";
 export * from "./lib/enum";
 export * from "./lib/localStorage";
 export * from "./lib/string";
+export * from "./lib/guid";