+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from "./lib/date";
 export * from "./lib/enum";
 export * from "./lib/localStorage";
+export * from "./lib/string";
+export * from "./lib/guid";